streamlit_app/requirements.txt
streamlit
pandas
plotly
numpy
matplotlib
seaborn
scipy
kaleido
selenium
beautifulsoup4
webdriver-manager
pillow
ddddocr
